third_party/blink/web_tests/http/tests/websocket/cookie-http-to-ws-expected.txt

Test how WebSocket handles cookies with cookie attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cookie is "ws-path-origin-script=1; ws-path-root-domain-local-ip=1"
PASS cookie is "ws-path-root-domain-local-ip=1"
PASS successfullyParsed is true

TEST COMPLETE
